refactor(architect-web): relocate ProtocolCard to ProtocolSummary feature
legacy ProtocolCard had one consumer and many unused props. alllows for easier removal of BEM markers
