+++ apps/architect-web/src/lib/ProtocolSummary/components/Cover.tsx
@@ -1,7 +1,7 @@
 import { DateTime } from "luxon";
 import { useContext } from "react";
 import networkCanvasLogo from "~/images/NC-Mark.svg";
-import ProtocolCard from "~/lib/legacy-ui/components/Cards/ProtocolCard";
+import ProtocolCard from "./ProtocolCard";
 import SummaryContext from "./SummaryContext";
 
 const Cover = () => {
@@ -14,7 +14,7 @@ const Cover = () => {
 	const now = `${date.toLocaleDateString()} ${date.toLocaleTimeString()}`;
 
 	return (
-		<div className="relative flex h-(--page-size-height) flex-col items-center justify-center print:h-screen [&_.protocol-card]:bg-platinum [&_.protocol-card]:max-w-[12cm] [&_.protocol-card]:[zoom:120%] [&_.protocol-card_.protocol-name]:block">
+		<div className="relative flex h-(--page-size-height) flex-col items-center justify-center print:h-screen">
 			<div className="absolute top-0 left-0 flex w-full items-center justify-between border-b-2 border-platinum">
 				<h2>Protocol Summary Document</h2>
 				<div className="flex items-center justify-end">
